move get_user_choice function to a new file
