+++ src/SConscript
@@ -585,6 +585,7 @@ wesnoth_sources = Split("""
     synced_checkup.cpp
     synced_context.cpp
     synced_commands.cpp
+    synced_user_choice.cpp
     syncmp_handler.cpp
     team.cpp
     teambuilder.cpp
