+++ src/synced_checkup.cpp
@@ -18,6 +18,7 @@
 #include "unit_map.hpp"
 #include "unit.hpp"
 #include "replay.hpp"
+#include "synced_user_choice.hpp"
 #include "resources.hpp"
 #include "game_display.hpp"
 static lg::log_domain log_replay("replay");
