+++ src/wesnoth.cpp
@@ -43,7 +43,6 @@
 #include "network.hpp"			// for describe_versions
 #include "preferences.hpp"              // for core_id, etc
 #include "preferences_display.hpp"      // for display_manager
-#include "replay.hpp"                   // for recorder, replay
 #include "scripting/application_lua_kernel.hpp"
 #include "scripting/plugins/context.hpp"
 #include "scripting/plugins/manager.hpp"
